boards/TinyFPGA_B2: Setup bootloader build for board
Using the TinyFPGA_BX as a base, setup the constraints and module
definition to build the bootloader bitstream for the TinyFPGA B2. All
build artifacts are populated into a build/ subdirectory to simplify
`make clean` and .gitignore setup.

The `fw.bin` targets builds a complete SPI flash image. It includes the
metadata section in flash for "bootmeta", however it leaves the security
page setup as an external process/step.

Also included as preliminary Make targets to build this design with
nextpnr instead of arachne-pnr. The firmware image can be build with the
follow command `make build/bootloader.nextpnr.img`.