tinyprog: Pad to full minor sector write size
Either the TinyProg-BootLoader state machines, or the SPI flash chip
used on the TinyFPGA BX, seems unable to *properly* program a partial
minor sector (ie, writing less than 256 octets), leading to a failed
write/verify cycle.  This seems to happen at the end of any programming
that is not an exact multiple of 256 octets.  Since the whole 256 octets
is already erased in program_sectors(), we pad out the short write to
a full 256 octets (with 0xff, which is the value read for "erased", to
reduce wear on on the flash cells).

For more detail on diagnosing this, see:
https://github.com/timvideos/litex-buildenv/issues/137